Accessing the database of epiworld

Accessing the database of epiworld

get_hist_total(x)

get_transition_probability(x)

get_reproductive_number(x)

# S3 method for epiworld_repnum
plot(
  x,
  y = NULL,
  ylab = "Average Rep. Number",
  xlab = "Day (step)",
  main = "Daily Average Reproductive Number",
  ...
)

get_hist_transition_matrix(x, skip_zeros = FALSE)

# S3 method for epiworld_hist_transition
as.array(x, ...)

Arguments

x

An object of class epiworld_sir, epiworld_seir, etc. any model.

...

In the case of plot.epiworld_repnum, further arguments passed to graphics::plot.

skip_zeros

Logical scalar. When FALSE it will return all the entries in the transition matrix.

Value

In the case of get_hist_transition_matrix, a data.frame with four columns: "state_from", "state_to", "date", and "counts." It will be

The as.array method for epiworld_hist_transition objects turns the data.frame

returned by get_hist_transition_matrix into an array of nstates x nstates x (ndays + 1)

entries, where the first entry is the initial state.

See also

Examples

# SEIR Connected
seirconn <- ModelSEIRCONN(
  name              = "Disease",
  n                 = 10000,
  prevalence        = 0.1,
  contact_rate      = 2.0, 
  prob_transmission = 0.8,
  incubation_days   = 7.0,
  prob_recovery     = 0.3
)

# Running the simulation for 50 steps (days)
set.seed(937)
run(seirconn, 50)
#> _________________________________________________________________________
#> |Running the model...
#> |||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||| done.
#> | done.

# Retrieving the transition probability
get_transition_probability(seirconn) 
#>             Susceptible   Exposed  Infected Recovered
#> Susceptible   0.8997312 0.1002688 0.0000000 0.0000000
#> Exposed       0.0000000 0.8582575 0.1417426 0.0000000
#> Infected      0.0000000 0.0000000 0.6964839 0.3035161
#> Recovered     0.0000000 0.0000000 0.0000000 1.0000000

# We can go further and get all the history
t_hist <- get_hist_transition_matrix(seirconn)

head(t_hist)
#>    state_from    state_to date counts
#> 1 Susceptible Susceptible    0   9000
#> 2     Exposed Susceptible    0      0
#> 3    Infected Susceptible    0      0
#> 4   Recovered Susceptible    0      0
#> 5 Susceptible     Exposed    0   1000
#> 6     Exposed     Exposed    0      0

# And turn it into an array
as.array(t_hist)[,,1:3]
#> , , 0
#> 
#>             Susceptible Exposed Infected Recovered
#> Susceptible        9000    1000        0         0
#> Exposed               0       0        0         0
#> Infected              0       0        0         0
#> Recovered             0       0        0         0
#> 
#> , , 1
#> 
#>             Susceptible Exposed Infected Recovered
#> Susceptible        9000       0        0         0
#> Exposed               0     861      139         0
#> Infected              0       0        0         0
#> Recovered             0       0        0         0
#> 
#> , , 2
#> 
#>             Susceptible Exposed Infected Recovered
#> Susceptible        8820     180        0         0
#> Exposed               0     737      124         0
#> Infected              0       0       91        48
#> Recovered             0       0        0         0
#>