Virus

Virus

virus(name, post_immunity, prob_infecting, prob_recovery, prob_death)

add_virus(m, v, prevalence)

add_virus_n(m, v, prevalence)

virus_set_state(v, init, end, removed)

rm_virus(m, virus_pos)

Arguments

name

of the virus

post_immunity, prob_infecting, prob_recovery, prob_death

to be documented

m

An object of class epiworld-model.

v

An object of class epiworld_virus

prevalence

In the case of add_virus, a proportion, otherwise, an integer.

init, end, removed

states after acquiring a virus, removing a virus, and removing the agent as a result of the virus, respectively.

virus_pos

Positive integer. Index of the virus's position in the model.

Examples

mseirconn <- ModelSEIRCONN(
  name                = "COVID-19",
  prevalence          = 0.01, 
  n                   = 10000,
  contact_rate        = 4, 
  incubation_days     = 7, 
  prob_transmission   = 0.5,
  prob_recovery       = 0.99
)

delta <- virus("Delta Variant", 0, .5, .2, .01)
virus_set_state(delta, 1, 3, 3)

add_virus(mseirconn, delta, .3)

run(mseirconn, ndays = 100, seed = 992)
#> _________________________________________________________________________
#> Running the model...
#> ||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||||| done.
#>  done.
mseirconn
#> 
#> ________________________________________________________________________________
#> SIMULATION STUDY
#> 
#> Name of the model   : Susceptible-Exposed-Infected-Removed (SEIR) (connected)
#> Population size     : 10000
#> Number of entities  : 0
#> Days (duration)     : 100 (of 100)
#> Number of variants  : 2
#> Last run elapsed t  : 40.00ms
#> Last run speed      : 24.91 million agents x day / second
#> Rewiring            : off
#> 
#> Virus(es):
#>  - COVID-19 (baseline prevalence: 1.00%)
#>  - Delta Variant (baseline prevalence: 30.00%)
#> 
#> Tool(s):
#>  (none)
#> 
#> Model parameters:
#>  - Avg. Incubation days : 7.0000
#>  - Contact rate         : 4.0000
#>  - Prob. Recovery       : 0.9900
#>  - Prob. Transmission   : 0.5000
#> 
#> Distribution of the population at time 100:
#>  - (0) Susceptible :  6900 -> 1205
#>  - (1) Exposed     :  3100 -> 0
#>  - (2) Infected    :     0 -> 0
#>  - (3) Recovered   :     0 -> 8795
#> 
#> Transition Probabilities:
#>  - Susceptible  0.98  0.02  0.00  0.00
#>  - Exposed      0.00  0.84  0.16  0.00
#>  - Infected     0.00  0.00  0.01  0.99
#>  - Recovered    0.00  0.00  0.00  1.00
#>